Fix unnecessary conversion warnings

+++ consensus/ethash/algorithm.go
@@ -304,11 +304,11 @@ func generateDataset(dest []uint32, epoch uint64, cache []uint32) {
 			keccak512 := makeHasher(sha3.NewLegacyKeccak512())
 
 			// Calculate the data segment this thread should generate
-			batch := uint64((size + hashBytes*uint64(threads) - 1) / (hashBytes * uint64(threads)))
+			batch := (size + hashBytes*uint64(threads) - 1) / (hashBytes * uint64(threads))
 			first := uint64(id) * batch
 			limit := first + batch
-			if limit > uint64(size/hashBytes) {
-				limit = uint64(size / hashBytes)
+			if limit > size/hashBytes {
+				limit = size / hashBytes
 			}
 			// Calculate the dataset segment
 			percent := uint32(size / hashBytes / 100)
